+++ web/indexPage.dart
@@ -22,6 +22,7 @@ void main() {
     apps.add(new DemoPage("desktop", new LIconUtility(LIconUtility.DESKTOP), "Desktop"));
     apps.add(new DemoFrame("tablet", new LIconUtility(LIconUtility.TABLET_PORTRAIT), "Tablet", 768));
     apps.add(new DemoFrame("phone", new LIconUtility(LIconUtility.PHONE_PORTRAIT), "Phone", 480));
+    apps.add(new ExampleForm());
     page.set(apps);
   });